+++ front/srcs/styles.css
@@ -6,18 +6,22 @@ body {
     min-width: 500px;
     font-family: 'SCoreDream';
     font-weight: 500;
+    height: 100vh;
 }
 
-html,
-body,
-#app,
-.page
+#app {
+  width: 100%;
+  height: 100%;
+}
+
+.page 
 {
   margin: 0;
   padding: 0;
   overflow: hidden;
-  width: 100vw;
-  height: 100vh;
+  width: 100%;
+  height: 100%;
+  display: block;
 }
 
 .canvas_container {
@@ -32,7 +36,6 @@ body,
   outline: none;
 }
 
-.home-view {
 .page-name {
     color: #7D9F56;
     text-shadow: 2px 1px 0 rgba(0, 0, 0, 0.154);
