+++ front/srcs/views/game/game_view.html
@@ -9,6 +9,7 @@
       width: 100%;
     .logo-image {
       width: 150px;
+      margin-top: 0.2em;
       height: auto;
     }
     #start-button{
