--- front/srcs/views/viewClasses.js
@@ -1,34 +1,0 @@
-import GameView from "@/views/game/game_view";
-import LoginView from "@/views/login/login_view";
-import HomeView from "@/views/home/home_view";
-import FriendView from "@/views/friend/friend_view";
-import ModeView from "@/views/mode/mode_view";
-import RecordView from "@/views/record/record_view";
-import TournamentView from "@/views/tournament/tournament_view";
-import MatchView from "@/views/match/match_view";
-import NavBar from "@/views/components/nav_bar";
-import ProfileCard from "@/views/components/profile_card";
-
-export default {
-  GameView,
-  HomeView,
-  LoginView,
-  FriendView,
-  ModeView,
-  RecordView,
-  TournamentView,
-  MatchView,
-  NavBar,
-  ProfileCard,
-};
-
-export const routes = [
-  { path: "/", view: HomeView},
-  { path: "/login", view: LoginView},
-  { path: "/friend", view: FriendView},
-  { path: "/record", view: RecordView},
-  { path: "/game", view: GameView },
-  { path: "/mode", view: ModeView },
-  { path: "/match", view: MatchView },
-  { path: "/tournament", view: TournamentView },
-];
